chore: 컴포넌트 파일명 명사형으로 변경
