+++ src/components/common/modal/TeamMemberInviteModal.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import styled from 'styled-components';
 import InputField from '@/components/common/InputField';
 import Button from '@/components/common/button/Button';
@@ -51,74 +51,55 @@ const S = {
       justify-content: space-between;
     }
   `,
-  DeleteButton: styled.button`
-    position: absolute;
-    bottom: 2rem;
-    left: 2rem;
-
-    display: flex;
-    justify-content: flex-start;
-
-    color: ${({ theme }) => theme.color.gray};
-    font-size: 1.4rem;
-    text-decoration-line: underline;
-
-    ${MEDIA_QUERIES.onMobile} {
-      margin-bottom: 1.6rem;
-    }
-  `,
 };
 
-interface ManageColumnsModalProps {
+interface TeamMemberInviteModalProps {
   isOpen: boolean;
   onClose: () => void;
-  onDelete: () => void;
-  onChange: (columnName: string) => void;
-  currentColumnName: string;
+  onCreate: (email: string) => void;
 }
 
-function ManageColumnsModal({
+function TeamMemberInviteModal({
   isOpen,
   onClose,
-  onChange,
-  onDelete,
-  currentColumnName = '',
-}: ManageColumnsModalProps) {
-  const [columnName, setColumnName] = useState(currentColumnName);
+  onCreate,
+}: TeamMemberInviteModalProps) {
+  const [email, setColumnName] = useState('');
 
-  useEffect(() => {
-    setColumnName(currentColumnName);
-  }, [currentColumnName]);
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setColumnName(e.target.value);
   };
 
-  const handleChange = () => {
-    onChange(columnName);
+  const handleClose = () => {
+    setColumnName('');
     onClose();
   };
 
+  const handleEmail = () => {
+    onCreate(email);
+    handleClose();
+  };
+
   return (
-    <BackDropModal isOpen={isOpen} onClose={onClose}>
-      <S.Title>컬럼 관리</S.Title>
+    <BackDropModal isOpen={isOpen} onClose={handleClose}>
+      <S.Title>초대하기</S.Title>
       <S.Input
         label="이름"
-        id="manageColumn"
-        placeholder="새로운 프로젝트"
-        value={columnName}
+        id="inviteTeamMember"
+        placeholder="codeit@codeit.com"
+        value={email}
         onChange={handleInputChange}
       />
-      <S.DeleteButton onClick={onDelete}>삭제하기</S.DeleteButton>
       <S.ButtonContainer>
         <S.ImportButton onClick={onClose} styleType="SECONDARY" size="M">
           취소
         </S.ImportButton>
-        <S.ImportButton onClick={handleChange} styleType="PRIMARY" size="M">
-          생성
+        <S.ImportButton onClick={handleEmail} styleType="PRIMARY" size="M">
+          초대
         </S.ImportButton>
       </S.ButtonContainer>
     </BackDropModal>
   );
 }
 
-export default ManageColumnsModal;
+export default TeamMemberInviteModal;
